refactor(sdk): split plexi_sdk/__init__.py into focused modules
Break 2980-line monolith into: _constants, _types, _emitter, _pipe,
_render_context, _app. __init__.py becomes a thin explicit re-export
layer. Public import surface unchanged — all examples continue to work
with no modifications.
